// src/Api/models/create-blog-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Hey Pharamcist API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface CreateBlogDto
 */
export interface CreateBlogDto {
  _id?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateBlogDto
     */
    blogTitle: string;
    /**
     * 
     * @type {string}
     * @memberof CreateBlogDto
     */
    image: string;
    /**
     * 
     * @type {string}
     * @memberof CreateBlogDto
     */
    content: string;
    /**
     * 
     * @type {boolean}
     * @memberof CreateBlogDto
     */
    isActive: boolean;
}